Improved login functionality

+++ client/controllers/CRUDCtrl.js
@@ -7,23 +7,33 @@ CRUDController.$inject = ['$http', '$rootScope', '$routeParams', 'UserService', 
 function CRUDController($http, $rootScope, $routeParams, userService, attractionService) { 
   
   var vm = this;
+  vm.isLoggedIn = $rootScope.isLoggedIn;
   
   //Hämtar en inloggad användare
-  var userPromise = userService.getUser($rootScope.user_id);
-  userPromise.then(function(data){
-    vm.username = data.username;
-    vm.attractions = data.attractions;
-  }).catch(function(error){
-    vm.message = error;
-  })
-  
-  //Tar bort en turistattraktion
-  vm.deleteAttraction = function(id) {
-    var deletePromise = attractionService.deleteAttraction(id);
-    deletePromise.then(function(data){
-      vm.message = "The tourist attraction has been deleted.";
+  if($rootScope.isLoggedIn){
+    var userPromise = userService.getUser($rootScope.user_id);
+    userPromise.then(function(data){
+      vm.username = data.username;
+      vm.attractions = data.attractions;
     }).catch(function(error){
       vm.message = error;
     })
-  };
+  } else {
+    vm.message = "You are not logged in.";
+  }
+  
+  //Tar bort en turistattraktion
+  if($rootScope.isLoggedIn){
+    vm.deleteAttraction = function(id) {
+      var deletePromise = attractionService.deleteAttraction(id);
+      deletePromise.then(function(data){
+        vm.message = "The tourist attraction has been deleted.";
+      }).catch(function(error){
+        vm.message = error;
+      })
+    }
+  } else {
+    vm.message = "You are not logged in.";
+  }
+  
 }
